// types/squareupSheetLine.ts
export interface Product {
  referenceHandle: string;
  token: string;
  itemName: string;
  variationName: string;
  unitAndPrecision: string;
  sku: string;
  description: string;
  reportingCategory: string;
  seoTitle: string;
  seoDescription: string;
  permalink: string;
  gtin: string;
  squareOnlineItemVisibility: string;
  itemType: string;
  weight: string;
  socialMediaLinkTitle: string;
  socialMediaLinkDescription: string;
  shippingEnabled: string;
  selfServeOrderingEnabled: string;
  deliveryEnabled: string;
  pickupEnabled: string;
  price: number;
  onlineSalePrice: string;
  sellable: string;
  stockable: string;
  skipDetailScreenInPOS: string;
  optionName1: string;
  optionValue1: string;
  optionName2: string;
  optionValue2: string;
  enabledBeyondParadiseShop: string;
  currentQuantityBeyondParadiseShop: number;
  newQuantityBeyondParadiseShop: string;
  stockAlertEnabledBeyondParadiseShop: string;
  stockAlertCountBeyondParadiseShop: string;
  priceBeyondParadiseShop: string;
  enabledMindfulImageriesGallery: string;
  currentQuantityMindfulImageriesGallery: number;
  newQuantityMindfulImageriesGallery: string;
  stockAlertEnabledMindfulImageriesGallery: string;
  stockAlertCountMindfulImageriesGallery: string;
  priceMindfulImageriesGallery: string;
}